requirements.txt
django==4.2.2
psycopg2-binary 
djangorestframework